Fix StyleDelve (damage, evade/parry/block req)
Add StyleDelve class
Move body from DetailDisplayHandler.DelveStyle to StyleDelve
Replace MiniDelveWriter usage with ClientDelve
Add logic for damage classification (None, Low, Medium, etc)
Add evade, parry, block requirement in StyleDelve
