+++ GameServer/spells/SpellDelve.cs
@@ -17,10 +17,7 @@
  *
  */
 using DOL.GS.PacketHandler;
-using DOL.GS.Styles;
-using System;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace DOL.GS.Spells
 {
@@ -133,80 +130,4 @@ namespace DOL.GS.Spells
 			return 0;
 		}
 	}
-
-	public class StyleDelve
-    {
-		private GameClient clt;
-		private int id;
-		private Style style;
-
-		private short TooltipId => unchecked((short)id);
-		private int LevelZeroDamage => (int)Math.Round(style.GrowthOffset / 0.295);
-		private int DamageIncreasePerLevel => (int)Math.Round(style.GrowthRate / 0.295);
-
-		public StyleDelve(GameClient clt, int id)
-        {
-			this.clt = clt;
-			this.id = id;
-			var sk = clt.Player.GetAllUsableSkills().Where(e => e.Item1.InternalID == id && e.Item1 is Style).FirstOrDefault();
-
-			if (sk == null || sk.Item1 == null)
-			{
-				style = SkillBase.GetStyleByInternalID(id);
-			}
-			else if (sk.Item1 is Style)
-			{
-				style = (Style)sk.Item1;
-			}
-		}
-
-		public string GetClientMessage()
-        {
-			if (style != null)
-			{
-				var styles = clt.Player.GetSpecList().SelectMany(e => e.PretendStylesForLiving(clt.Player, clt.Player.MaxLevel));
-
-				var clientDelve = new ClientDelve("Style");
-				clientDelve.AddElement("Index", TooltipId);
-
-				if (style.OpeningRequirementType == Style.eOpening.Offensive && style.AttackResultRequirement == Style.eAttackResultRequirement.Style)
-				{
-					Style st = styles.Where(s => s.ID == style.OpeningRequirementValue).FirstOrDefault();
-					if (st != null)
-					{
-						clientDelve.AddElement("OpeningStyle", st.Name);
-					}
-				}
-
-				var followupStyles = styles
-					.Where(s => (s.OpeningRequirementType == Style.eOpening.Offensive && s.AttackResultRequirement == Style.eAttackResultRequirement.Style && s.OpeningRequirementValue == style.ID))
-					.Select(s => s.Name);
-				clientDelve.AddElement("FollowupStyle", followupStyles);
-				clientDelve.AddElement("Name", style.Name);
-				clientDelve.AddElement("Icon", style.Icon);
-				clientDelve.AddElement("Level", style.Level);
-				clientDelve.AddElement("Fatigue", style.EnduranceCost);
-				clientDelve.AddElement("DefensiveMod", style.BonusToDefense);
-				clientDelve.AddElement("AttackMod", style.BonusToHit);
-				clientDelve.AddElement("OpeningDamage", LevelZeroDamage + style.Level * DamageIncreasePerLevel);
-				clientDelve.AddElement("LevelBonus", DamageIncreasePerLevel);
-				clientDelve.AddElement("OpeningType", (int)style.OpeningRequirementType);
-				if (style.OpeningRequirementType == Style.eOpening.Positional)
-					clientDelve.AddElement("OpeningNumber", style.OpeningRequirementValue);
-				if (style.WeaponTypeRequirement > 0)
-					clientDelve.AddElement("Weapon", style.GetRequiredWeaponName());
-				clientDelve.AddElement("OpeningResult", (int)style.AttackResultRequirement);
-				clientDelve.AddElement("Hidden", style.StealthRequirement);
-
-				return clientDelve.ClientMessage;
-			}
-			else
-			{
-				var clientDelve = new ClientDelve("Style");
-				clientDelve.AddElement("Index", TooltipId);
-				clientDelve.AddElement("Name", "(not found)");
-				return clientDelve.ClientMessage;
-			}
-		}
-    }
 }
